Split out TokenRenderer file
