+++ browser/src/Renderer/DOMRenderer.ts
@@ -1,26 +1,16 @@
 import { IDeltaCellPosition, IDeltaRegionTracker } from "./../DeltaRegionTracker"
 import { Grid } from "./../Grid"
 import * as Performance from "./../Performance"
-import { ICell, IScreen } from "./../Screen"
+import { IScreen } from "./../Screen"
 import { INeovimRenderer } from "./INeovimRenderer"
 
 import { IElementFactory, RecycleElementFactory } from "./DOM/ElementFactory"
 
-import { ISpan, ISpanElementInfo, combineSpansAtBoundary, collapseSpanMap,} from "./DOM/Spans"
+import { ISpan, ISpanElementInfo, combineSpansAtBoundary, collapseSpanMap, } from "./DOM/Spans"
+
+import { ITokenRenderer, getRendererForCell } from "./DOM/TokenRenderer"
 
 // TODO: Look at scroll perf
-
-export interface ITokenRenderer {
-    x: number
-    y: number
-    width: number
-    backgroundColor: string | undefined
-    foregroundColor: string | undefined
-
-    canHandleCell(cell: ICell): boolean
-    appendCell(cell: ICell): void
-    getTag(): HTMLElement | null
-}
 
 export class DOMRenderer implements INeovimRenderer {
     private _editorElement: HTMLDivElement
@@ -126,165 +116,6 @@ export class DOMRenderer implements INeovimRenderer {
             this._grid.setCell(x, y, infoToSave)
         }
     }
-}
-
-export class BaseTokenRenderer {
-
-    private _screen: IScreen
-    private _backgroundColor: string | undefined
-    private _foregroundColor: string | undefined
-    private _x: number
-    private _y: number
-    private _width: number = 0
-    private _elementFactory: IElementFactory
-
-    private _lastCell: ICell // TODO: This is just used to work around TS compiler... Specifically, it doesn't like `cell` being passedin appendCell to be overridden
-
-    public get x(): number {
-        return this._x
-    }
-
-    public get y(): number {
-        return this._y
-    }
-
-    public get width(): number {
-        return this._width
-    }
-
-    public get foregroundColor(): string | undefined {
-        return this._foregroundColor
-    }
-
-    public get backgroundColor(): string | undefined {
-        return this._backgroundColor
-    }
-
-    protected get screen(): IScreen {
-        return this._screen
-    }
-
-    constructor(x: number, y: number, cell: ICell, screen: IScreen, elementFactory: IElementFactory) {
-        this._foregroundColor = cell.foregroundColor
-        this._backgroundColor = cell.backgroundColor
-        this._screen = screen
-        this._x = x
-        this._y = y
-        this._elementFactory = elementFactory
-    }
-
-    public canHandleCell(cell: ICell): boolean {
-        return this._foregroundColor === cell.foregroundColor && this._backgroundColor === cell.backgroundColor
-    }
-
-    public appendCell(cell: ICell): void {
-        this._lastCell = cell
-
-        this._width++
-    }
-
-    public getDefaultTag(): HTMLElement {
-
-        const span = this._elementFactory.getElement()
-        span.style.position = "absolute"
-        span.style.top = (this._y * this.screen.fontHeightInPixels) + "px"
-        span.style.left = (this._x * this.screen.fontWidthInPixels) + "px"
-        span.style.height = this.screen.fontHeightInPixels + "px"
-
-        if (this._backgroundColor && this._backgroundColor !== this._screen.backgroundColor) {
-            span.style.backgroundColor = this._backgroundColor
-        }
-
-        span.style.color = this._foregroundColor || this._screen.foregroundColor
-        return span
-
-    }
-}
-
-export class WhiteSpaceTokenRenderer extends BaseTokenRenderer implements ITokenRenderer {
-
-    constructor(x: number, y: number, cell: ICell, screen: IScreen, elementFactory: IElementFactory) {
-        super(x, y, cell, screen, elementFactory)
-    }
-
-    public canHandleCell(cell: ICell): boolean {
-        return super.canHandleCell(cell) && isWhiteSpace(cell)
-    }
-
-    public getTag(): HTMLElement | null {
-        if (!this.backgroundColor || this.backgroundColor === this.screen.backgroundColor) {
-            return null
-        }
-
-        const span = super.getDefaultTag()
-        span.className = "whitespace"
-        span.style.width = (this.width * this.screen.fontWidthInPixels) + "px"
-        return span
-    }
-}
-
-export class TokenRenderer extends BaseTokenRenderer implements ITokenRenderer {
-
-    private _str: string = ""
-
-    constructor(x: number, y: number, cell: ICell, screen: IScreen, elementFactory: IElementFactory) {
-        super(x, y, cell, screen, elementFactory)
-    }
-
-    public canHandleCell(cell: ICell): boolean {
-        return super.canHandleCell(cell) && !isWhiteSpace(cell)
-    }
-
-    public appendCell(cell: ICell): void {
-        super.appendCell(cell)
-
-        if (cell.characterWidth > 0) {
-            this._str += cell.character
-        }
-    }
-
-    public getTag(): HTMLElement | null {
-        const span = super.getDefaultTag()
-        span.textContent = this._str
-        span.style.width = ((this._str.length + 1) * this.screen.fontWidthInPixels) + "px"
-        return span
-    }
-
-}
-
-export function getRendererForCell(x: number, y: number, cell: ICell, screen: IScreen, elementFactory: IElementFactory) {
-    if (isWhiteSpace(cell)) {
-        return new WhiteSpaceTokenRenderer(x, y, cell, screen, elementFactory)
-    } else {
-        return new TokenRenderer(x, y, cell, screen, elementFactory)
-    }
-}
-
-export function isWhiteSpace(cell: ICell): boolean {
-    const character = cell.character
-
-    return cell.characterWidth === 1 && (character === " " || character === "" || character === "\t" || character === "\n")
-}
-
-export function addOrCoalesceSpan(existingSpans: ISpan[], newSpan: ISpan): ISpan[] {
-    const overlappingSpans = existingSpans.filter((s) => {
-
-        if ((newSpan.startX >= s.startX && newSpan.startX <= s.endX)
-            || (newSpan.endX >= s.startX && newSpan.endX <= s.endX)) {
-            return true
-        } else {
-            return false
-        }
-    })
-
-    const nonOverlappingSpans = existingSpans.filter((s) => overlappingSpans.indexOf(s) === -1)
-
-    const combinedSpan = overlappingSpans.reduce((prev, cur) => ({
-        startX: Math.min(prev.startX, cur.startX),
-        endX: Math.max(prev.endX, cur.endX),
-    }), newSpan)
-
-    return nonOverlappingSpans.concat([combinedSpan])
 }
 
 export function getSpansToEdit(grid: Grid<ISpanElementInfo>, cells: IDeltaCellPosition[], elementFactory: IElementFactory): Map<number, ISpan[]> {
